draft investigation icons
