+++ frontend/src/constants/engineConst.js
@@ -1,0 +1,14 @@
+export const EvaluationIcons = Object.freeze({
+  trusted: "like",
+  malicious: "malware",
+});
+
+export const TagsIcons = Object.freeze({
+  phishing: "hook",
+  anonymizer: "incognito",
+  malware: "malware",
+  social_engineering: "",
+  tor_exit_node: "tor",
+  scanner: "magnifyingGlass",
+  ip_only: "networkNode",
+});
