+++ frontend/tests/components/jobs/result/visualizer/elements/bool.test.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "@testing-library/jest-dom";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { getIcon } from "../../../../../../src/components/jobs/result/visualizer/icons";
+import { getIcon } from "../../../../../../src/components/common/icon/icons";
 import { BooleanVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/bool";
 
 // mock useLocation
