+++ frontend/tests/components/jobs/result/visualizer/icons.test.jsx
@@ -1,6 +1,6 @@
 const {
   getIcon,
-} = require("../../../../../src/components/jobs/result/visualizer/icons");
+} = require("../../../../../src/components/common/icon/icons");
 
 describe("visualizer icon module", () => {
   test("check defined icons", () => {
